Refactor domUpdates
