+++ src/scripts.js
@@ -11,6 +11,7 @@ import Recipe from './classes/Recipe';
 import User from './classes/User';
 import { userData, recipesData, ingredientData } from './apiCalls';
 import Pantry from './classes/Pantry';
+import domUpdates from './domUpdates';
 
 
 
@@ -55,25 +56,40 @@ let pantry;
 
 // fetch apis *************************************
 Promise.all([userData, recipesData, ingredientData])
-  .then(data => {
-    recipes = data[1].map(recipe => {
-      return new Recipe(recipe.id, recipe.image, recipe.ingredients, recipe.instructions, recipe.name, recipe.tags, data[2])
+  .then((data) => {
+    recipes = data[1].map((recipe) => {
+      return new Recipe(
+        recipe.id,
+        recipe.image,
+        recipe.ingredients,
+        recipe.instructions,
+        recipe.name,
+        recipe.tags,
+        data[2]
+      );
     });
-    users = data[0].map(user => {
-      return new User(user.name, user.id, pantry = new Pantry(user.pantry, data[2]), data[2])
+    users = data[0].map((user) => {
+      return new User(
+        user.name,
+        user.id,
+        (pantry = new Pantry(user.pantry, data[2])),
+        data[2]
+      );
     });
-  
 
     cookBook = new Cookbook(recipes);
     filter = cookBook;
     currentUser = users[Math.floor(Math.random() * users.length)];
-    currentUser.pantry.listIngredientsNameAndAmount()
-    displayRecipes();
+    currentUser.pantry.listIngredientsNameAndAmount();
+    domUpdates.displayRecipes(
+      singleRecipeSection,
+      favoriteRecipeSection,
+      allRecipesSection,
+      cookBook
+    );
     filterByCheckBoxes();
- 
-
   })
-  .catch(error => console.log('Oooops. Something is wrong', error))
+  .catch((error) => domUpdates.displayError(error, pageTitle));
 
 //reusable functions ***************************
 const hide = (element) => {
@@ -86,216 +102,7 @@ const show = (element) => {
 
 
 // Event Hnadlers *******************************
-const displayRecipes = () => {
-  hide(singleRecipeSection)
-  hide(favoriteRecipeSection)
-  const allRecipies = cookBook.recipes.forEach(recipe => {
-    return allRecipesSection.innerHTML += `
-      <div class='recipe' id='${recipe.id}' tabindex="0" >
-      <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-      <div class= 'name-and-favorite'>
-      <p class='recipe-name'>${recipe.name}</p>
-      </div>
-      </div>`
-  });
-  return allRecipies;
-}; 
 
-const showFavoriteRecipes = () => {
-  favoriteRecipeSection.innerHTML = ''
-  hide(allRecipesSection)
-  hide(singleRecipeSection)
-  filter = currentUser
-  pageTitle.innerHTML = 'Your Favorite Recipes';
-  const displayFavoriteRecipes = currentUser.favoriteRecipes.forEach(recipe => {
-    return favoriteRecipeSection.innerHTML += `
-    <div class='recipe' id='${recipe.id}' tabindex="0">
-        <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-        <h5>${recipe.name}</h5>
-        </div>` 
-  });
-  return displayFavoriteRecipes
-};
-
-const populateSearch = (e) => {
-  let input = e.target.value
-  if (input && input.trim().length > 0) {
-    input = input.trim().toLowerCase()
-    if (filter === cookBook) {
-      allRecipesSection.innerHTML = filter.filterByKeyWord(input).reduce((acc, recipe) => {
-        acc += `
-              <div class='recipe' id='${recipe.id}' tabindex="0">
-              <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-              <h5>${recipe.name}</h5>
-              </div> 
-              `
-        return acc
-      }, '')
-    } else {
-      favoriteRecipeSection.innerHTML = filter.filterByKeyWord(input).reduce((acc, recipe) => {
-        acc += `
-        <div class='recipe' id='${recipe.id} tabindex="0">
-        <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-        <h5>${recipe.name}</h5>
-        </div> 
-        `
-        return acc
-      }, '')
-    }
-  }
-};
-
-const clickRecipe = (event) => {
-  singleRecipeSection.innerHTML = ''
-  if (event.target.id) {
-    hide(favoriteRecipeSection)
-    hide(allRecipesSection)
-    hide(mealsToCookSection)
-    show(singleRecipeSection)
-    hide(searchBox)
-    hide(clearButton)
-    const findRecipeId = recipes.find(({ id }) => id == event.target.id)
-    currentRecipe = findRecipeId
-    const recipeInstructions = findRecipeId.instructions.reduce((acc, instruction) => {
-      acc += `<li>${instruction.instruction}</li>`
-      return acc
-    }, '')
-    currentRecipe.listIngredients()
-    const ingredientList = currentRecipe.listOfRecipeIngredients.reduce((acc, ingredient) => {
-      acc += `<li>${ingredient.name} Amount: ${ingredient.amount}</li>`
-      return acc
-    }, '');
-    singleRecipeSection.innerHTML = `
-      <div class="single-recipe-img" tabindex="0"> <img src="${findRecipeId.image}" alt=""> </div>
-      <div class="favorite-meal-buttons">
-        <button class='favorite-button' id="favorite-button"><img class="favorite-image" id="favorite-button" src="./images/like.png"  alt="favorite"> </button>
-        <button class='add-meal-button' id="add-meal-button"><img class="add-image" src="./images/plus.png" id="add-meal-button" alt="add"> </button>
-        <div class="total-cost"> <h1 class="price">$${findRecipeId.costOfIngredients()}</h1></div> 
-      </div>
-      <div class="ingredient-instructions">
-        <div class="ingredient-instructions-section">${recipeInstructions} </div>
-        <div class="ingredient-list"><p> ${ingredientList} </p> </div>
-      </div>
-    `
-  } else {
-    console.log('clicking outside')
-  }
-};
-
-
-const mealsToCookSingleRecipe = (event) => {
-  singleRecipeSection.innerHTML = ''
-  if (event.target.id) {
-    hide(favoriteRecipeSection)
-    hide(allRecipesSection)
-    hide(mealsToCookSection)
-    show(singleRecipeSection)
-    const findRecipeId = recipes.find(({ id }) => id == event.target.id)
-    currentRecipe = findRecipeId
-    console.log(currentRecipe,'Current recipe')
-    const recipeInstructions = findRecipeId.instructions.reduce((acc, instruction) => {
-      acc += `<li>${instruction.instruction}</li>`
-      return acc
-    }, '')
-    currentRecipe.listIngredients()
-    currentUser.pantry.checkPantry(currentRecipe.listOfRecipeIngredients)
-    const ingredientList = findRecipeId.listOfRecipeIngredients.reduce((acc, ingredient) => {
-      acc += `<li>${ingredient.name} Amount: ${ingredient.amount}</li>`
-      return acc
-    }, '');
-    const missingIngredientsList = currentUser.pantry.listOfMissingIngredients.reduce((acc,ingredient) => {
-      acc += `<li>${ingredient.name} Amount: ${ingredient.amount[0]}</li>`
-      return acc
-    },'')
-    console.log(currentUser.pantry.listOfMissingIngredients)
-    const pantryList = currentUser.pantry.listOfPantryIngredients.reduce((acc,ingredient) => {
-
-      if(ingredient.amount !== 0){
-        acc += `
-          <li>${ingredient.name} Amount: ${ingredient.amount}</li>
-      `
-      }
-      
-      return acc      
-    })
-
-    singleRecipeSection.innerHTML = `
-      <div class="single-recipe-img" > <img src="${findRecipeId.image}" alt=""> </div>
-      <div class="favorite-meal-buttons" tabindex="0">
-        <button class='shopping-cart-button'><img class="shopping-cart-img" id="shoppingCart"src="./images/shopping-cart.png" alt="shopping Cart"> </button>
-        <button class='cook-button'><img class="cook-img" src="./images/cooking.png" id="cookImg" alt="Cooking Img"> </button>
-        <div class="total-cost"> <h1 class="price">$${findRecipeId.costOfIngredients()}</h1></div> 
-      </div>
-      <div class="ingredient-instructions">
-        <div class="ingredient-instructions-section">${recipeInstructions} </div>
-        <div class="ingredient-list"><p> ${ingredientList} </p> </div>
-      </div>
-      <h1>***MISSING INGREDIENTS***</h1>
-      <p> ${missingIngredientsList} </p>
-
-      <h1>***PANTRY***</h1>
-
-      <div class='show-pantry'>
-      <p>${pantryList}</p>
-      </div>
-      
-    `
-  } else {
-    console.log('clicking outside')
-  }
-}
-
-const showFilteredRecipes = () => {
-  const filteredRecipes = filter.filterByTags(tagList).reduce((acc, recipe) => {
-    allRecipesSection.innerHTML = ``
-    pageTitle.innerHTML = `Filtered Recipes`
-    acc += `
-    <div class='recipe' id='${recipe.id}' tabindex="0">
-    <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-    <h5>${recipe.name}</h5>
-    </div>
-    `
-    return acc
-  }, '');
-  allRecipesSection.innerHTML = filteredRecipes
-};
-
-const showMealPlan = () => {
-  pageTitle.innerHTML = 'Meals to cook'
-  mealsToCookSection.innerHTML = '';
-  hide(allRecipesSection)
-  hide(singleRecipeSection)
-  show(mealsToCookSection)
-  hide(pantrySection)
-  hide(favoriteRecipeSection)
-  hide(searchBox)
-  filter = currentUser
-  const displayMealsToCook = currentUser.recipesToCook.forEach(recipe => {
-    return mealsToCookSection.innerHTML += `
-    <div class='recipe' id='${recipe.id}' tabindex="0">
-        <img class="recipe-image" src="${recipe.image}" id='${recipe.id}' alt="${recipe.name}">
-        <h5>${recipe.name}</h5>
-        </div>`
-  });
-  return displayMealsToCook
-};
-
-const showPantrySection = (section) => {
-  section.innerHTML = ``
-  
-  const pantryList = currentUser.pantry.listOfPantryIngredients.forEach(ingredient => {
-    if(ingredient.amount !== 0){
-      section.innerHTML += `
-      <tr class='ingredient-table'>
-        <td>${ingredient.name}</td>
-        <td>${ingredient.amount} units</td>
-      </tr>
-    `
-    }
-    
-  })
-  return pantryList
-}
 
 
 const addMissingIngredients = (event) => {
@@ -309,7 +116,7 @@ const cookMeal = (event) => {
     const showPantryDiv = document.querySelector('.show-pantry') 
     currentUser.pantry.cookRecipe(currentUser)
     currentUser.pantry.checkPantry(currentRecipe.listOfRecipeIngredients)
-    showPantrySection(showPantryDiv)
+    domUpdates.showPantrySection(showPantryDiv, currentUser);
   } else {
     console.log('Not enough ingredients');
   }
@@ -324,7 +131,7 @@ const filterByCheckBoxes = () => {
         tagList = Array.from(checkboxes)
           .filter(i => i.checked)
           .map(i => i.id)
-        showFilteredRecipes()
+        domUpdates.showFilteredRecipes(filter, allRecipesSection, pageTitle, tagList);
         checkboxes.forEach(checkbox => {
           if (!checkbox.checked) {
             checkbox.disabled = true;
@@ -333,12 +140,24 @@ const filterByCheckBoxes = () => {
       } else if (!this.checked && filter === cookBook) {
         cookBook.filteredRecipes = []
         allRecipesSection.innerHTML = ''
-        displayRecipes()
+        domUpdates.displayRecipes(
+          singleRecipeSection,
+          favoriteRecipeSection,
+          allRecipesSection,
+          cookBook
+        );
         checkboxes.forEach(checkbox => {
           checkbox.disabled = false
         })
       } else {
-        showFavoriteRecipes()
+        domUpdates.showFavoriteRecipes(
+          favoriteRecipeSection,
+          allRecipesSection,
+          singleRecipeSection,
+          currentUser,
+          filter,
+          pageTitle
+        );
         checkboxes.forEach(checkbox => {
           checkbox.disabled = false
         })
@@ -360,42 +179,111 @@ const addSingleRecipe = (event) => {
 }
 
 
+
+
   
 // event listeners ***********************************************
 
 allRecipesSection.addEventListener('click', (event) => {
-  clickRecipe(event)
+  domUpdates.clickRecipe(
+    event,
+    singleRecipeSection,
+    favoriteRecipeSection,
+    allRecipesSection,
+    mealsToCookSection,
+    searchBox,
+    clearButton,
+    currentRecipe,
+    recipes
+  );
 });
 
 allRecipesSection.addEventListener('keydown', (event) => {
   if (event.which === 13) {
-    clickRecipe(event)
+    domUpdates.clickRecipe(
+      event,
+      singleRecipeSection,
+      favoriteRecipeSection,
+      allRecipesSection,
+      mealsToCookSection,
+      searchBox,
+      clearButton,
+      currentRecipe,
+      recipes
+    );
   } else {
     console.log('outside')
   }
 });   
 favoriteRecipeSection.addEventListener('click', (event) => {
-  clickRecipe(event)
+  domUpdates.clickRecipe(
+    event,
+    singleRecipeSection,
+    favoriteRecipeSection,
+    allRecipesSection,
+    mealsToCookSection,
+    searchBox,
+    clearButton,
+    currentRecipe,
+    recipes
+  );
 });
 
 favoriteRecipeSection.addEventListener('keydown', (event) => {
   if (event.which === 13) {
-    clickRecipe(event)
+    domUpdates.clickRecipe(
+      event,
+      singleRecipeSection,
+      favoriteRecipeSection,
+      allRecipesSection,
+      mealsToCookSection,
+      searchBox,
+      clearButton,
+      currentRecipe,
+      recipes
+    );
   } else {
     console.log('outside')
   }
 });
  
-mealsToCookButton.addEventListener('click', showMealPlan);
+mealsToCookButton.addEventListener('click', () => {
+  domUpdates.showMealPlan(
+    pageTitle,
+    mealsToCookSection,
+    allRecipesSection,
+    pantrySection,
+    favoriteRecipeSection,
+    searchBox,
+    filter,
+    currentUser,
+    singleRecipeSection
+  );
+});
 
 
 mealsToCookSection.addEventListener('click', (event) => {
-  mealsToCookSingleRecipe(event)
+  domUpdates.mealsToCookSingleRecipe(
+    event,
+    singleRecipeSection,
+    favoriteRecipeSection,
+    allRecipesSection,
+    mealsToCookSection,
+    currentRecipe,
+    currentUser
+  );
   
 });
 
 favoriteRecipesButton.addEventListener("click", () => {
-  showFavoriteRecipes()
+  domUpdates.showFavoriteRecipes(
+    favoriteRecipeSection,
+    allRecipesSection,
+    singleRecipeSection,
+    currentUser,
+    filter,
+    pageTitle
+  );
   hide(clearButton)
   hide(pantrySection)
   hide(allRecipesSection)
@@ -403,12 +291,10 @@ favoriteRecipesButton.addEventListener("click", () => {
   show(favoriteRecipeSection)
   hide(searchBox)
   hide(tagNavSection)
-  
-  console.log('click')
 });
 
 searchBox.addEventListener('input', (e) => {
-  populateSearch(e)
+  domUpdates.populateSearch(e, filter, cookBook, allRecipesSection, favoriteRecipeSection);
 });
 
 clearButton.addEventListener("click", () => {
@@ -425,7 +311,12 @@ allRecipesButton.addEventListener('click', () => {
   hide(singleRecipeSection)
   hide(favoriteRecipeSection)
   show(allRecipesSection)
-  displayRecipes()
+  domUpdates.displayRecipes(
+    singleRecipeSection,
+    favoriteRecipeSection,
+    allRecipesSection,
+    cookBook
+  );
   show(clearButton)
   show(searchBox)
   hide(tagNavSection)
@@ -443,12 +334,20 @@ pantryButton.addEventListener('click', () => {
   hide(searchBox)
   hide(tagNavSection)
   show(pantrySection)
-  showPantrySection(pantryTable)
+  domUpdates.showPantrySection(pantryTable, currentUser);
 });
 
 mealsToCookSection.addEventListener('keydown', (event) => {
   if (event.which === 13) {
-    mealsToCookSingleRecipe(event)
+    domUpdates.mealsToCookSingleRecipe(
+      event,
+      singleRecipeSection,
+      favoriteRecipeSection,
+      allRecipesSection,
+      mealsToCookSection,
+      currentRecipe,
+      currentUser
+    );
   } else {
     console.log('outside')
   }
